Changed the GenericPartitioner to be specific to the RsuLogKey object
