+++ jpo-geojsonconverter/src/main/java/us/dot/its/jpo/geojsonconverter/partitioner/RsuLogKeyPartitioner.java
@@ -1,0 +1,30 @@
+package us.dot.its.jpo.geojsonconverter.partitioner;
+
+import org.apache.kafka.common.serialization.Serdes;
+import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.streams.processor.StreamPartitioner;
+
+import us.dot.its.jpo.geojsonconverter.serialization.serializers.JsonSerializer;
+
+public class RsuLogKeyPartitioner<K, V> implements StreamPartitioner<K, V> {
+    @Override
+    public Integer partition(String topic, K key, V value, int numPartitions) {
+        byte[] partitionBytes;
+        
+        if (key instanceof RsuLogKey) {
+            // If the key is an object with an RSU ID, partition on it
+            var rsuIdKey = (RsuLogKey)key;
+            String bsmId = rsuIdKey.getBsmId();
+            try (var serializer = Serdes.String().serializer()) {
+                partitionBytes = Serdes.String().serializer().serialize(topic, bsmId); 
+            }          
+        } else {
+            // If the key does not have an RSU ID, partition on the hashed key as usual.
+            try (var serializer = new JsonSerializer<K>()) {
+                partitionBytes = serializer.serialize(topic, key);
+            }
+        }
+
+        return Utils.toPositive(Utils.murmur2(partitionBytes)) % numPartitions;
+    }
+}
