+++ jpo-geojsonconverter/src/test/java/us/dot/its/jpo/geojsonconverter/partitioner/RsuLogKeyPartitionerTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertNotEquals;
 
 import org.junit.Test;
 
-public class GenericPartitionerTest {
+public class RsuLogKeyPartitionerTest {
     @Test
     public void testPartition_RsuLogKey() {
 
@@ -13,12 +13,12 @@ public class GenericPartitionerTest {
         
         final String topic = "topic";
         RsuLogKey key = new RsuLogKey("127.0.0.1", null, "ABCDEFG");
-        RsuLogKey sameKey = new RsuLogKey("127.0.0.1", null, "ABCDEFG");
-        RsuLogKey differentKey = new RsuLogKey(null, "bsmTx.gz", "ABCDEFG");
+        RsuLogKey sameKey = new RsuLogKey(null, null, "ABCDEFG");
+        RsuLogKey differentKey = new RsuLogKey(null, "bsmTx.gz", "GFEDCBA");
         final Object obj = new Object();
         final int numPartitions = Integer.MAX_VALUE;
 
-        var partitioner = new GenericPartitioner<RsuLogKey, Object>();
+        var partitioner = new RsuLogKeyPartitioner<RsuLogKey, Object>();
 
         int partitionKey = partitioner.partition(topic, key, obj, numPartitions);
         int partitionSame = partitioner.partition(topic, sameKey, obj, numPartitions);
@@ -40,7 +40,7 @@ public class GenericPartitionerTest {
         final Object obj = new Object();
         final int numPartitions = Integer.MAX_VALUE;
 
-        var partitioner = new GenericPartitioner<String, Object>();
+        var partitioner = new RsuLogKeyPartitioner<String, Object>();
 
         int partitionKey = partitioner.partition(topic, key, obj, numPartitions);
         int partitionSame = partitioner.partition(topic, sameKey, obj, numPartitions);
